finished/tested: - paladin - retribution Aura - rage for spell dmg - warlock: Contagion - warlock: Nightfall. Shadow trance is not removed by first shadow bolt yet. New exploit :P - warlock: Improved Searing Pain - warlock - Shadow Mastery - warlock - Improved Curse of Agony - warlock - Grim Reach - warlock - Suppression - warlock: Empowered Corruption - warlock: Malediction - warlock - Shadow and Flame - warlock - Emberstorm - warlock: intensity - warlock: Devastation
- DOTS use the same spell dmg bonus calculation algorithm as normal spells do. This needs some wider testing then i was able to make in a few hours


git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2502 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/MailSystem.cpp
@@ -70,7 +70,7 @@ WorldPacket * Mailbox::BuildMailboxListingPacket()
 		if(itr->second.expire_time && t > itr->second.expire_time)
 			continue;	   // expired mail -> skip it
 
-		if(UNIXTIME < itr->second.delivery_time)
+		if((uint32)UNIXTIME < itr->second.delivery_time)
 			continue;		// undelivered
 		
 		if(itr->second.AddMessageDataToPacket(*data))
@@ -780,7 +780,7 @@ void Mailbox::FillTimePacket(WorldPacket& data)
 
 	for(; iter != Messages.end(); ++iter)
 	{
-		if(iter->second.read_flag == 0 && UNIXTIME >= iter->second.delivery_time)
+		if(iter->second.read_flag == 0 && (uint32)UNIXTIME >= iter->second.delivery_time)
 		{
 			// unread message, w00t.
 			++c;
